LayoutTests/http/tests/security/contentSecurityPolicy/extensions/manifest-v3/default-src/manifest-v3-default-src-block-non-https-scheme-expected.txt

ALERT: PASS successfully blocked blob-URL script.

Tests CSP for manifest v3 extensions does not permit non https schemes
